refactor: setup everything after merge
