+++ src/main/java/org/storkforge/barkr/config/Config.java
@@ -2,15 +2,15 @@ package org.storkforge.barkr.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.storkforge.barkr.domain.entity.Account;
 import org.storkforge.barkr.domain.entity.Post;
-import org.storkforge.barkr.web.infrastructure.persistence.AccountRepository;
+import org.springframework.boot.CommandLineRunner;
+import org.storkforge.barkr.domain.entity.Account;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Profile;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
 import org.storkforge.barkr.web.infrastructure.persistence.PostRepository;
+import org.storkforge.barkr.web.infrastructure.persistence.AccountRepository;
 
 import java.util.List;
