+++ src/main/java/org/storkforge/barkr/domain/services/api/AccountService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.storkforge.barkr.domain.entity.Account;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 import org.storkforge.barkr.exceptions.AccountNotFound;
-import org.storkforge.barkr.infrastructure.persistence.AccountRepository;
+import org.storkforge.barkr.web.infrastructure.persistence.AccountRepository;
 import org.storkforge.barkr.mapper.AccountMapper;
 
 import java.util.List;
