+++ src/main/java/org/storkforge/barkr/dto/accountDto/ResponseAccount.java
@@ -1,9 +1,6 @@
 package org.storkforge.barkr.dto.accountDto;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.PastOrPresent;
-import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.*;
 
 import java.time.LocalDateTime;
