+++ src/main/java/org/storkforge/barkr/web/graphql/AccountResolver.java
@@ -1,4 +1,4 @@
-package org.storkforge.barkr.graphql;
+package org.storkforge.barkr.web.graphql;
 
 
 
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Positive;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
-import org.storkforge.barkr.domain.AccountService;
+import org.storkforge.barkr.web.domain.AccountService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 
 import java.util.List;
